% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{Environment}
\alias{Environment}
\title{Environment}
\description{
An R6 class that represents a Content's Environment Variables
}
\section{Super class}{
\code{\link[connectapiR6:Content]{connectapiR6::Content}} -> \code{Environment}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Environment$new()}}
\item \href{#method-env_refresh}{\code{Environment$env_refresh()}}
\item \href{#method-print}{\code{Environment$print()}}
\item \href{#method-clone}{\code{Environment$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="environment">}\href{../../connectapiR6/html/Content.html#method-environment}{\code{connectapiR6::Content$environment()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="environment_set">}\href{../../connectapiR6/html/Content.html#method-environment_set}{\code{connectapiR6::Content$environment_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_bundles">}\href{../../connectapiR6/html/Content.html#method-get_bundles}{\code{connectapiR6::Content$get_bundles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_connect">}\href{../../connectapiR6/html/Content.html#method-get_connect}{\code{connectapiR6::Content$get_connect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_content">}\href{../../connectapiR6/html/Content.html#method-get_content}{\code{connectapiR6::Content$get_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_content_remote">}\href{../../connectapiR6/html/Content.html#method-get_content_remote}{\code{connectapiR6::Content$get_content_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_dashboard_url">}\href{../../connectapiR6/html/Content.html#method-get_dashboard_url}{\code{connectapiR6::Content$get_dashboard_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_job">}\href{../../connectapiR6/html/Content.html#method-get_job}{\code{connectapiR6::Content$get_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_jobs">}\href{../../connectapiR6/html/Content.html#method-get_jobs}{\code{connectapiR6::Content$get_jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_url">}\href{../../connectapiR6/html/Content.html#method-get_url}{\code{connectapiR6::Content$get_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="job">}\href{../../connectapiR6/html/Content.html#method-job}{\code{connectapiR6::Content$job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="jobs">}\href{../../connectapiR6/html/Content.html#method-jobs}{\code{connectapiR6::Content$jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="runas">}\href{../../connectapiR6/html/Content.html#method-runas}{\code{connectapiR6::Content$runas()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="tag_delete">}\href{../../connectapiR6/html/Content.html#method-tag_delete}{\code{connectapiR6::Content$tag_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="tag_set">}\href{../../connectapiR6/html/Content.html#method-tag_set}{\code{connectapiR6::Content$tag_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="tags">}\href{../../connectapiR6/html/Content.html#method-tags}{\code{connectapiR6::Content$tags()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="update">}\href{../../connectapiR6/html/Content.html#method-update}{\code{connectapiR6::Content$update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="variants">}\href{../../connectapiR6/html/Content.html#method-variants}{\code{connectapiR6::Content$variants()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Environment$new(connect, content)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-env_refresh"></a>}}
\if{latex}{\out{\hypertarget{method-env_refresh}{}}}
\subsection{Method \code{env_refresh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Environment$env_refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Environment$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Environment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
