% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant.R
\name{VariantTask}
\alias{VariantTask}
\title{VariantTask}
\description{
An R6 class that represents a Variant Task
}
\seealso{
Other R6 classes: 
\code{\link{Content}},
\code{\link{RStudioConnect}}
}
\concept{R6 classes}
\section{Super classes}{
\code{\link[connectapiR6:Content]{connectapiR6::Content}} -> \code{\link[connectapiR6:Variant]{connectapiR6::Variant}} -> \code{VariantTask}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{VariantTask$new()}}
\item \href{#method-get_task}{\code{VariantTask$get_task()}}
\item \href{#method-print}{\code{VariantTask$print()}}
\item \href{#method-clone}{\code{VariantTask$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="environment">}\href{../../connectapiR6/html/Content.html#method-environment}{\code{connectapiR6::Content$environment()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="environment_set">}\href{../../connectapiR6/html/Content.html#method-environment_set}{\code{connectapiR6::Content$environment_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_bundles">}\href{../../connectapiR6/html/Content.html#method-get_bundles}{\code{connectapiR6::Content$get_bundles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_connect">}\href{../../connectapiR6/html/Content.html#method-get_connect}{\code{connectapiR6::Content$get_connect()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_content">}\href{../../connectapiR6/html/Content.html#method-get_content}{\code{connectapiR6::Content$get_content()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_content_remote">}\href{../../connectapiR6/html/Content.html#method-get_content_remote}{\code{connectapiR6::Content$get_content_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_job">}\href{../../connectapiR6/html/Content.html#method-get_job}{\code{connectapiR6::Content$get_job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="get_jobs">}\href{../../connectapiR6/html/Content.html#method-get_jobs}{\code{connectapiR6::Content$get_jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="runas">}\href{../../connectapiR6/html/Content.html#method-runas}{\code{connectapiR6::Content$runas()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="tag_delete">}\href{../../connectapiR6/html/Content.html#method-tag_delete}{\code{connectapiR6::Content$tag_delete()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="tag_set">}\href{../../connectapiR6/html/Content.html#method-tag_set}{\code{connectapiR6::Content$tag_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="tags">}\href{../../connectapiR6/html/Content.html#method-tags}{\code{connectapiR6::Content$tags()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="update">}\href{../../connectapiR6/html/Content.html#method-update}{\code{connectapiR6::Content$update()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Content" data-id="variants">}\href{../../connectapiR6/html/Content.html#method-variants}{\code{connectapiR6::Content$variants()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="add_subscribers">}\href{../../connectapiR6/html/Variant.html#method-add_subscribers}{\code{connectapiR6::Variant$add_subscribers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="get_dashboard_url">}\href{../../connectapiR6/html/Variant.html#method-get_dashboard_url}{\code{connectapiR6::Variant$get_dashboard_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="get_schedule">}\href{../../connectapiR6/html/Variant.html#method-get_schedule}{\code{connectapiR6::Variant$get_schedule()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="get_schedule_remote">}\href{../../connectapiR6/html/Variant.html#method-get_schedule_remote}{\code{connectapiR6::Variant$get_schedule_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="get_subscribers">}\href{../../connectapiR6/html/Variant.html#method-get_subscribers}{\code{connectapiR6::Variant$get_subscribers()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="get_url">}\href{../../connectapiR6/html/Variant.html#method-get_url}{\code{connectapiR6::Variant$get_url()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="get_url_rev">}\href{../../connectapiR6/html/Variant.html#method-get_url_rev}{\code{connectapiR6::Variant$get_url_rev()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="get_variant">}\href{../../connectapiR6/html/Variant.html#method-get_variant}{\code{connectapiR6::Variant$get_variant()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="get_variant_remote">}\href{../../connectapiR6/html/Variant.html#method-get_variant_remote}{\code{connectapiR6::Variant$get_variant_remote()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="job">}\href{../../connectapiR6/html/Variant.html#method-job}{\code{connectapiR6::Variant$job()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="jobs">}\href{../../connectapiR6/html/Variant.html#method-jobs}{\code{connectapiR6::Variant$jobs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="remove_subscriber">}\href{../../connectapiR6/html/Variant.html#method-remove_subscriber}{\code{connectapiR6::Variant$remove_subscriber()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="render">}\href{../../connectapiR6/html/Variant.html#method-render}{\code{connectapiR6::Variant$render()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="renderings">}\href{../../connectapiR6/html/Variant.html#method-renderings}{\code{connectapiR6::Variant$renderings()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="send_mail">}\href{../../connectapiR6/html/Variant.html#method-send_mail}{\code{connectapiR6::Variant$send_mail()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="connectapiR6" data-topic="Variant" data-id="update_variant">}\href{../../connectapiR6/html/Variant.html#method-update_variant}{\code{connectapiR6::Variant$update_variant()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$new(connect, content, key, task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_task"></a>}}
\if{latex}{\out{\hypertarget{method-get_task}{}}}
\subsection{Method \code{get_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$get_task()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
